third_party/WebKit/LayoutTests/http/tests/htmlimports/cross-origin-expected.txt

CONSOLE ERROR: Imported resource from origin 'http://localhost:8080' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
PASS target.import is null
PASS successfullyParsed is true

TEST COMPLETE
